feat: Remove duplicate code
