+++ packages/blog/src/utils/transformArticle.ts
@@ -1,25 +1,4 @@
-import { PaginationType, ResponseArticleDataType } from '../types'
-
-export interface ArticleDataType {
-  id: number
-  slug: string
-  title: string
-  locale: string
-  imgUrl: string
-  content: string
-  createAt: string
-  publishedAt: string
-  description: string
-  categories: Array<string>
-  newsOutBoundLink: string
-  newsFromPlatform: string
-  gamesCategories: Array<string>
-}
-
-export interface ArticleType {
-  data: ArticleDataType[]
-  pagination: PaginationType
-}
+import { ArticleDataType, ResponseArticleDataType } from '../types'
 
 export const transformArticle = (article: ResponseArticleDataType): ArticleDataType => {
   return {
